+++ CHANGELOG.md
@@ -27,6 +27,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Fixed bug in PageLayout with `wide` prop.
 
+- Updated sort icons for table.
+
+- IconHelper components group is separated into a discrete entity.
+
 ## [0.21.0] - 2020-08-11
 
 - Fix crashes in Chromium when checkbox was clicked fast three times in a row.